pyproject.toml
[tool.poetry]
name = "telegram_bot_zoo"
version = "0.1.0"
description = ""
authors = ["Gmacem <gmacem@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.0"
sqlalchemy-stubs = "^0.4"
mypy = "^0.991"
python-dotenv = "^0.21.1"
loguru = "^0.6.0"
aiohttp = "^3.8.3"
fastapi = "^0.89.1"
alembic = "^1.9.2"
psycopg = "^3.1.8"
dependency-injector = "^4.41.0"
yarl = "^1.8.2"
click = "^8.1.3"
greenlet = "^2.0.2"
uvicorn = "^0.20.0"
python-telegram-bot = "^20.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ruff = "^0.0.241"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
